requirements.txt
# requirements.txt
psycopg2-binary
pytest
requests
allure-pytest
locust
playwright
